repository pattern get!

+++ TurtleAPI/AlphaVantage/AlphaVMarketDetail.cs
@@ -26,19 +26,19 @@ namespace TurtleAPI.AlphaVantage
         [JsonProperty("01. symbol")]
         public string? symbol { get; set; }
         [JsonProperty("02. open")]
-        public float open { get; set; }
+        public decimal open { get; set; }
         [JsonProperty("03. high")]
-        public float high { get; set; }
+        public decimal high { get; set; }
         [JsonProperty("04. low")]
-        public float low { get; set; }
+        public decimal low { get; set; }
         [JsonProperty("05. price")]
-        public float price { get; set; }
+        public decimal price { get; set; }
         [JsonProperty("06. volume")]
-        public float volume { get; set; }
+        public long volume { get; set; }
         [JsonProperty("07. latest trading day")]
         public DateTime latest { get; set; }
         [JsonProperty("08. previous close")]
-        public float previous { get; set; }
+        public decimal previous { get; set; }
         
     }
 }
